chore: move autoScrollingMarquee into design-system-hedvig
Promote Modifier.autoScrollingMarquee() from feature-onboarding to the shared
design-system-hedvig commonMain so any surface can reuse it. Now public and
KMP-common (available to iOS too).
